Correção no pom

+++ src/main/resources/db/migration/V1_000__create_cliente.sql
@@ -1,5 +1,5 @@
 CREATE TABLE if not exists CLIENTE (
-  id long NOT NULL AUTO_INCREMENT,
-  nome varchar(200) NOT NULL
+  id integer NOT NULL AUTO_INCREMENT,
+  nome varchar(200) NOT NULL,
   PRIMARY KEY (id)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
